cmake: avaota-a1: allow CROSS_COMPILE environment to set cross compiler
Many projects (including the Linux kernel and U-Boot) expect the cross
compiler triplet prefix to be specified in a shell environment variable
named CROSS_COMPILE. We seem to follow this, but require this prefix to
be specified on the cmake command line (-DCROSS_COMPILE=...).

Allow the prefix to be taken from the environment instead. This makes it
compatible with the build environment of those other projects.