avaota/extlinux: remove LCD code
The Avaota-A1 has a LCD display connected via SPI, which is used to
display boot status information. However this is of course only on this
board, other A523/T527 boards do not offer this display.

To make the code more versatile, remove the LCD code, so that other
boards can use the boot code as well.